#370 Set default DeleteBehavior for all relationship to Restrict

+++ src/Modules/SimplCommerce.Module.Core/wwwroot/admin/countries/country-list.js
@@ -32,14 +32,18 @@
         };
 
         vm.deleteCountry = function (country) {
-            countryService.deleteCountry(country)
-                .then(function (result) {
-                    vm.getCountries(tableStateRef);
-                    toastr.success(country.name + ' has been deleted');
-                })
-                .catch(function (response) {
-                    toastr.error(response.data.error);
-                });
+            bootbox.confirm('Are you sure you want to delete this country: ' + country.name, function (result) {
+                if (result) {
+                    countryService.deleteCountry(country)
+                        .then(function (result) {
+                            vm.getCountries(tableStateRef);
+                            toastr.success(country.name + ' has been deleted');
+                        })
+                        .catch(function (response) {
+                            toastr.error(response.data.error);
+                        });
+                }
+            });
         };
     }
 })();
